pyproject.toml
[tool.poetry]
name = "switchbot-cloud"
version = "0.1.0"
description = "A Python API client for controlling SwitchBot devices via the SwitchBot Cloud API"
authors = ["Eric Abruzzese <eric.abruzzese@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6.2"
pyhumps = "^3.0.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.11b1"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
